requirements.txt
Package                  Version
------------------------ -----------
annotated-doc            0.0.4
annotated-types          0.7.0
antlr4-python3-runtime   4.9.3
anyio                    4.12.1
asttokens                3.0.1
certifi                  2026.2.25
charset-normalizer       3.4.4
click                    8.3.1
comm                     0.2.3
contourpy                1.3.2
cuda-bindings            12.9.4
cuda-pathfinder          1.4.0
cycler                   0.12.1
debugpy                  1.8.20
decorator                5.2.1
exceptiongroup           1.3.1
executing                2.2.1
filelock                 3.24.3
fonttools                4.61.1
fsspec                   2026.2.0
gitdb                    4.0.12
gitpython                3.1.46
h11                      0.16.0
hf-xet                   1.3.2
httpcore                 1.0.9
httpx                    0.28.1
huggingface-hub          1.5.0
hydra-core               1.3.2
idna                     3.11
ipykernel                7.2.0
ipython                  8.38.0
jedi                     0.19.2
jinja2                   3.1.6
joblib                   1.5.3
jupyter-client           8.8.0
jupyter-core             5.9.1
kiwisolver               1.4.9
llvmlite                 0.46.0
markdown-it-py           4.0.0
markupsafe               3.0.3
matplotlib               3.10.8
matplotlib-inline        0.2.1
mdurl                    0.1.2
mpmath                   1.3.0
narwhals                 2.17.0
nest-asyncio             1.6.0
networkx                 3.4.2
numba                    0.64.0
numpy                    2.2.6
nvidia-cublas-cu12       12.8.4.1
nvidia-cuda-cupti-cu12   12.8.90
nvidia-cuda-nvrtc-cu12   12.8.93
nvidia-cuda-runtime-cu12 12.8.90
nvidia-cudnn-cu12        9.10.2.21
nvidia-cufft-cu12        11.3.3.83
nvidia-cufile-cu12       1.13.1.3
nvidia-curand-cu12       10.3.9.90
nvidia-cusolver-cu12     11.7.3.90
nvidia-cusparse-cu12     12.5.8.93
nvidia-cusparselt-cu12   0.7.1
nvidia-nccl-cu12         2.27.5
nvidia-nvjitlink-cu12    12.8.93
nvidia-nvshmem-cu12      3.4.5
nvidia-nvtx-cu12         12.8.90
omegaconf                2.3.0
opencv-python            4.13.0.92
packaging                26.0
pandas                   2.3.3
parso                    0.8.6
pexpect                  4.9.0
pillow                   12.1.1
pip                      26.0.1
platformdirs             4.9.4
plotly                   6.6.0
prompt-toolkit           3.0.52
protobuf                 6.33.5
psutil                   7.2.2
ptyprocess               0.7.0
pure-eval                0.2.3
pydantic                 2.12.5
pydantic-core            2.41.5
pygments                 2.19.2
pynndescent              0.6.0
pyparsing                3.3.2
python-dateutil          2.9.0.post0
python-math              0.0.1
pytz                     2025.2
pyyaml                   6.0.3
pyzmq                    27.1.0
regex                    2026.2.28
requests                 2.32.5
rich                     14.3.3
safetensors              0.7.0
scikit-learn             1.7.2
scipy                    1.15.3
sentry-sdk               2.53.0
shellingham              1.5.4
six                      1.17.0
smmap                    5.0.2
stack-data               0.6.3
sympy                    1.14.0
threadpoolctl            3.6.0
tokenizers               0.22.2
torch                    2.10.0
torchvision              0.25.0
tornado                  6.5.4
tqdm                     4.67.3
traitlets                5.14.3
transformers             5.2.0
triton                   3.6.0
typer                    0.24.1
typer-slim               0.24.0
typing-extensions        4.15.0
typing-inspection        0.4.2
tzdata                   2025.3
umap-learn               0.5.11
urllib3                  2.6.3
wandb                    0.25.0
wcwidth                  0.6.0
